Configure Hystrix to use SEMAPHORES to avoid Spring transaction breakage
Hystrix uses thread isolation by default. The new thread does not
inherit Spring context or transaction status and all entities become
disconnected.

It is propably possible to fix this using new hystrix concurrency
strategy that would copy the context and transaction status over,
but I was not able to do this atm.