Change the default flush mode to COMMIT
Relinked object is in inconsistent state during the operation so any
flush will cause constraint violation.

The relinker is mixing queries and updates to the relinked object and
all queries generate flushes in the AUTO setting.

This change modifies hibernate default to use COMMIT that causes flush
operation only when the transaction is about to be commited.

Manual flush can be used in cases when a transient object is supposed
to be included in query results (relinker does not need this as it
maintains its own internal cache which is consulted first).